% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amr}
\alias{amr}
\title{Antimicrobial Resistance (AMR) Genes}
\format{
A data frame with 62791 rows and 34 variables:
\describe{
\item{Name}{Genome ID matching metadata id}
\item{Protein.identifier}{NCBI protein accessions}
\item{Contig.id}{Contig identifier where the gene is located}
\item{Start}{Start position in contig}
\item{Stop}{Stop position in contig}
\item{Strand}{Strand direction (+ or -)}
\item{Gene.symbol}{Symbol of the AMR gene (e.g. sdeA)}
\item{Sequence.name}{Detailed product sequence name}
\item{Scope}{Scope of resistance}
\item{Element.type}{Element type (e.g. AMR, STRESS)}
\item{Element.subtype}{Element subtype (e.g. BIOCIDE)}
\item{Class}{Drug resistance class}
\item{Subclass}{Drug resistance subclass}
\item{Method}{Method of detection}
\item{Target.length}{Target reference sequence length}
\item{Reference.sequence.length}{Length of reference sequence}
\item{X..Coverage.of.reference.sequence}{Percentage coverage}
\item{X..Identity.to.reference.sequence}{Percentage identity}
\item{Alignment.length}{Length of alignment}
\item{Accession.of.closest.sequence}{Accession number of closest sequence}
\item{Name.of.closest.sequence}{Name of closest sequence}
\item{HMM.id}{HMM identifier}
\item{HMM.description}{HMM model description}
\item{Hierarchy.node}{AMR ontology node}
\item{Protein.id}{Protein identifier}
\item{Element.symbol}{Element symbol}
\item{Element.name}{Element name}
\item{Type}{Type classification}
\item{Subtype}{Subtype classification}
\item{X..Coverage.of.reference}{Percentage coverage fallback}
\item{X..Identity.to.reference}{Percentage identity fallback}
\item{Closest.reference.accession}{Closest reference accession}
\item{Closest.reference.name}{Closest reference name}
\item{HMM.accession}{HMM accession}
}
}
\usage{
amr
}
\description{
A dataset containing AMR genes detected in various genomes.
}
\keyword{datasets}
