% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escape_xml}
\alias{escape_xml}
\title{Escape special XML characters}
\usage{
escape_xml(x)
}
\arguments{
\item{x}{A character vector to escape.}
}
\value{
A character vector with XML entities escaped.
}
\description{
Escape special XML characters
}
\keyword{internal}
