% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronar.R
\name{get_asset_path}
\alias{get_asset_path}
\title{Helper to resolve and read packaged assets}
\usage{
get_asset_path(filename)
}
\arguments{
\item{filename}{File name of the asset (e.g. "krona-2.0.js").}
}
\value{
File content as a single string.
}
\description{
Helper to resolve and read packaged assets
}
\keyword{internal}
