% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronar.R
\name{knit_print.kronar_plot}
\alias{knit_print.kronar_plot}
\title{Knit Print Method for Kronar Plot}
\usage{
\method{knit_print}{kronar_plot}(x, ...)
}
\arguments{
\item{x}{An object of class \code{kronar_plot}.}

\item{...}{Additional arguments passed to knit_print.}
}
\value{
Knitted HTML markup.
}
\description{
Knit Print Method for Kronar Plot
}
