% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata of Serratia Genomes}
\format{
A data frame with 3863 rows and 20 variables:
\describe{
\item{id}{Genome ID}
\item{query_signature}{Query signature hash}
\item{best_hit_id}{Identifier of the best hit genome}
\item{best_hit_signature}{Signature of the best hit}
\item{similarity_score}{ANI/similarity score to best hit}
\item{levels_reached}{Deepest taxonomic level resolved}
\item{final_code}{GTDB code}
\item{best_hit_code}{GTDB code of best hit}
\item{L_0}{Taxonomic rank level 0 (e.g. Kingdom)}
\item{L_1}{Taxonomic rank level 1 (e.g. Phylum)}
\item{L_2}{Taxonomic rank level 2 (e.g. Class)}
\item{L_3}{Taxonomic rank level 3 (e.g. Order)}
\item{L_4}{Taxonomic rank level 4 (e.g. Family)}
\item{L_5}{Taxonomic rank level 5 (e.g. Genus)}
\item{Assigned_Species}{Assigned species name}
\item{Relative_Abundance....}{Relative abundance percentage of the genome}
\item{ANI....}{ANI similarity percentage}
\item{gtdb_species}{GTDB assigned species name}
\item{Colection}{Collection category}
\item{Assigned_Species2}{Secondary species assignment}
}
}
\usage{
metadata
}
\description{
A dataset containing metadata for Serratia genomes, including taxonomy levels
and relative abundances.
}
\keyword{datasets}
