% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronar.R
\name{print.kronar_plot}
\alias{print.kronar_plot}
\title{Print Kronar Plot}
\usage{
\method{print}{kronar_plot}(x, ...)
}
\arguments{
\item{x}{An object of class \code{kronar_plot}.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
The original object, invisibly.
}
\description{
Automatically displays the Krona plot in a web browser or viewer if run in
interactive mode, or prints the HTML markup otherwise.
}
