% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_fill_colors}
\alias{resolve_fill_colors}
\title{Resolve fill colors from a column}
\usage{
resolve_fill_colors(df, fill_col, fill_palette = NULL)
}
\arguments{
\item{df}{Data frame.}

\item{fill_col}{Name or index of the column containing fill values/colors.}

\item{fill_palette}{Optional custom palette. For numeric columns, a vector of colors defining a gradient. For discrete columns, a vector of colors or name of a palette.}
}
\value{
A character vector of hexadecimal colors of the same length as rows in \code{df}.
}
\description{
Resolve fill colors from a column
}
\keyword{internal}
