% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_and_parse_df}
\alias{validate_and_parse_df}
\title{Validate and parse the input data frame for KronaR}
\usage{
validate_and_parse_df(df, count_col = NULL, fill_col = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{count_col}{Name or index of the column containing counts.}

\item{fill_col}{Name or index of the column containing fill values/colors.}
}
\value{
A list containing \code{hier} (a data frame of character columns) and \code{counts} (a numeric vector).
}
\description{
Validate and parse the input data frame for KronaR
}
\keyword{internal}
